Crate pallet_runtime_configs

Pallet for tweaking the runtime configs for multiple network.

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

weights

Autogenerated weights for pallet_runtime_configs

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Traits

Config

Configuration trait of this pallet.

Type Aliases

ConfirmationDepthK

Storage type is [StorageValue] with value type BlockNumberFor < T >.

CouncilDemocracyConfig

Storage type is [StorageValue] with value type CouncilDemocracyConfigParams < BlockNumberFor < T > >.

EnableBalanceTransfers

Whether to enable balances transfers.

EnableDomains

Whether to enable calls in pallet-domains.

EnableDynamicCostOfStorage

Whether to enable dynamic cost of storage.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.