pallet_runtime_configs

Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

ConfirmationDepthK

Warning: Doc-Only

CouncilDemocracyConfig

Warning: Doc-Only

DomainBlockPruningDepth

Domains block pruning depth.

EnableBalanceTransfers

Whether to enable balances transfers.

EnableDomains

Whether to enable calls in pallet-domains.

EnableDynamicCostOfStorage

Whether to enable dynamic cost of storage.

StakingWithdrawalPeriod

Domain nominator’s staking withdrawal period